driver/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "monomod"
version = "0.1.0"
description = "Python driver for MONOMOD wireless EMG modules"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pyyaml",
]

[project.optional-dependencies]
gui = ["pyqt6", "pyqtgraph"]
lsl = ["pylsl"]
record = ["h5py"]
discovery = ["zeroconf"]

[tool.setuptools.packages.find]
where = ["src"]
